done_3

+++ funct_py_app.py
@@ -3,7 +3,7 @@ def average():
         try:
             lst = input("enter nums: ")
             lst=lst.split()
-            lst = [int(x) for x in lst]  
+            lst = [float(x) for x in lst]  
             result = sum(lst) / len(lst)
             print(result)
             return result
